Envi Template:
- change compressed file form tgz to zip

If we run uncompress of "tar" file, we have mistake output in terminal e.g.

- A lone zero block at 129

This issue occurs for certain versions of tar. The warning is apparently caused by different versions of tar having different end of file indicators. There isn't a definitive link on the topic, but please check out